requirements.txt
pytest
selene>=2.0.0.b17
allure-pytest